changing format folders
