+++ app/src/main/java/com/example/entregaandroidavanzadov2/AppClass.kt
@@ -2,8 +2,6 @@ package com.example.entregaandroidavanzadov2
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
-
-
 @HiltAndroidApp
 class AppClass: Application() {
 }
