+++ app/src/main/java/com/example/entregaandroidavanzadov2/LocationsHero.kt
@@ -1,9 +1,6 @@
 package com.example.entregaandroidavanzadov2
-
 data class LocationsHero(
 
     val latitud: Double,
     val longitud: Double,
-
-
 )
