+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/RepositoryLogin.kt
@@ -1,7 +1,5 @@
 package com.example.entregaandroidavanzadov2.data
 
 interface RepositoryLogin {
-
     suspend fun login(credentials: String): String?
-
 }
