+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/RepositoryLoginImpl.kt
@@ -6,18 +6,15 @@ import javax.inject.Inject
 
 class RepositoryLoginImpl @Inject constructor(
     private val remoteDataSource: RemoteDataSource
-): RepositoryLogin {
-
+    ): RepositoryLogin {
     override suspend fun login(credentials: String): String? {
         val response = remoteDataSource.login(credentials)
 
         if( response.isSuccessful){
             remoteDataSource.token = response.body().toString()
             return response.body()
-
         }else{
             return null
         }
-
     }
 }
