+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/local/HeroDatabase.kt
@@ -1,11 +1,8 @@
 package com.example.entregaandroidavanzadov2.data.local
 
-import androidx.room.AutoMigration
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import com.example.entregaandroidavanzadov2.data.local.model.LocalHero
-
-
 @Database(entities = [LocalHero::class], version = 1)
 abstract class HeroDatabase: RoomDatabase() {
