+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/local/model/LocalHero.kt
@@ -3,8 +3,6 @@ package com.example.entregaandroidavanzadov2.data.local.model
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-
-
 @Entity(tableName = "heros")
 data class LocalHero(
