+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/ILoginAPI.kt
@@ -5,8 +5,6 @@ import retrofit2.http.Header
 import retrofit2.http.POST
 
 interface ILoginAPI {
-
     @POST("api/auth/login")
     suspend fun login(@Header("Authorization") credentials: String): Response<String>
-
 }
