--- app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/Request/FavouriteHeroRequestBody.kt
@@ -1,7 +1,0 @@
-package com.example.entregaandroidavanzadov2.data.remote.Request
-
-data class FavouriteHeroRequestBody(
-
-    val hero: String
-
-)
