--- app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/Request/GetHeroLocationRequestBody.kt
@@ -1,6 +1,0 @@
-package com.example.entregaandroidavanzadov2.data.remote.Request
-
-data class GetHeroLocationRequestBody(
-
-    val id: String
-)
