--- app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/Request/GetHerosRequestBody.kt
@@ -1,7 +1,0 @@
-package com.example.entregaandroidavanzadov2.data.remote.Request
-
-data class GetHerosRequestBody(
-
-    val name: String = ""
-
-)
