+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/request/FavouriteHeroRequestBody.kt
@@ -1,0 +1,4 @@
+package com.example.entregaandroidavanzadov2.data.remote.request
+data class FavouriteHeroRequestBody(
+    val hero: String
+)
