+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/request/GetHeroLocationRequestBody.kt
@@ -1,0 +1,4 @@
+package com.example.entregaandroidavanzadov2.data.remote.request
+data class GetHeroLocationRequestBody(
+    val id: String
+)
