+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/request/GetHerosRequestBody.kt
@@ -1,0 +1,5 @@
+package com.example.entregaandroidavanzadov2.data.remote.request
+
+data class GetHerosRequestBody(
+    val name: String = ""
+)
