+++ app/src/main/java/com/example/entregaandroidavanzadov2/data/remote/response/GetHerosResponse.kt
@@ -1,15 +1,13 @@
-package com.example.entregaandroidavanzadov2.data.remote.Response
+package com.example.entregaandroidavanzadov2.data.remote.response
 
 
 import com.squareup.moshi.Json
 
 data class GetHerosResponse(
 
-
     @Json(name = "id") val id: String,
     @Json(name = "name") val name : String,
     @Json(name = "description") val description : String,
     @Json(name = "favorite") val favorite : Boolean,
-    @Json(name = "photo") val photo : String,
-
+    @Json(name = "photo") val photo : String
 )
