+++ app/src/main/java/com/example/entregaandroidavanzadov2/ui/HeroListFragment.kt
@@ -1,14 +1,12 @@
 package com.example.entregaandroidavanzadov2.ui
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import com.example.entregaandroidavanzadov2.R
 import com.example.entregaandroidavanzadov2.databinding.FragmentListBinding
 import com.example.entregaandroidavanzadov2.ui.viewModels.HerosViewModel
 import dagger.hilt.android.AndroidEntryPoint
@@ -20,12 +18,8 @@ import dagger.hilt.android.AndroidEntryPoint
 class HeroListFragment : Fragment() {
 
     private var _binding: FragmentListBinding? = null
-
-    // This property is only valid between onCreateView and
-    // onDestroyView.
     private val binding get() = _binding!!
     private val viewModel: HerosViewModel by viewModels()
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
     }
@@ -33,12 +27,10 @@ class HeroListFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
         _binding = FragmentListBinding.inflate(inflater, container, false)
         return binding.root
 
     }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -49,7 +41,6 @@ class HeroListFragment : Fragment() {
                 )
             )
         }
-
         binding.herosList.adapter = adapter
         viewModel.getHeros()
 
@@ -57,7 +48,6 @@ class HeroListFragment : Fragment() {
             adapter.submitList(heros)
         }
     }
-
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
