+++ app/src/main/java/com/example/entregaandroidavanzadov2/ui/SuperHeroAdapter.kt
@@ -14,36 +14,27 @@ import com.squareup.picasso.Picasso
 
 class SuperHeroAdapter(private val onClick: (String)-> (Unit)):
     ListAdapter<SuperHero, SuperHeroAdapter.SuperHeroViewHolder>(SuperHeroDiffCallBack()) {
-
-
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SuperHeroViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.hero_item, parent, false)
         return SuperHeroViewHolder(view)
     }
-
     override fun onBindViewHolder(holder: SuperHeroViewHolder, position: Int) {
         holder.bind(getItem(position))
     }
-
     override fun onViewRecycled(holder: SuperHeroViewHolder) {
         super.onViewRecycled(holder)
         holder.clear()
     }
-
     inner class SuperHeroViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
         private val heroName = itemView.findViewById<TextView>(R.id.hero_name)
         private val heroPhoto = itemView.findViewById<ImageView>(R.id.hero_photo)
-
         private lateinit var hero: SuperHero
-
-
         init {
             itemView.setOnClickListener {
                 onClick(hero.id)
             }
         }
-
         fun bind(hero: SuperHero) {
             this.hero = hero
             heroName.text = hero.name
@@ -54,23 +45,17 @@ class SuperHeroAdapter(private val onClick: (String)-> (Unit)):
                 .centerInside()
                 .into(heroPhoto)
         }
-
         fun clear() {
             Picasso.get().cancelRequest(heroPhoto)
             heroPhoto.setImageDrawable(null)
         }
-
-
     }
-
     class SuperHeroDiffCallBack : DiffUtil.ItemCallback<SuperHero>() {
         override fun areItemsTheSame(oldItem: SuperHero, newItem: SuperHero): Boolean {
             return oldItem.id == newItem.id
-
         }
         override fun areContentsTheSame(oldItem: SuperHero, newItem: SuperHero): Boolean {
             return oldItem == newItem
         }
-
     }
 }
