+++ app/src/main/java/com/example/entregaandroidavanzadov2/ui/viewModels/HerosViewModel.kt
@@ -4,24 +4,18 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.entregaandroidavanzadov2.LocationsHero
 import com.example.entregaandroidavanzadov2.SuperHero
 import com.example.entregaandroidavanzadov2.data.Repository
-import com.example.entregaandroidavanzadov2.data.remote.Response.GetHeroLocationResponse
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
-
 @HiltViewModel
 class HerosViewModel @Inject constructor(private val repository: Repository): ViewModel() {
 
-
     private val _heros = MutableLiveData<List<SuperHero>>()
     val heros: LiveData<List<SuperHero>> get() = _heros
-
-
     fun getHeros() {
         viewModelScope.launch {
             val result = withContext(Dispatchers.IO) {
@@ -30,5 +24,4 @@ class HerosViewModel @Inject constructor(private val repository: Repository): Vi
             _heros.value = result
         }
     }
-
 }
