+++ app/src/main/java/com/example/entregaandroidavanzadov2/utils/NetworkModule.kt
@@ -13,19 +13,15 @@ import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-
-
 @Module
 @InstallIn(SingletonComponent::class)
 object NetworkModule {
-
     @Provides
     fun providesMoshi(): Moshi{
         return Moshi.Builder()
             .addLast(KotlinJsonAdapterFactory())
             .build()
     }
-
     @Provides
     fun providesOkhttp(): OkHttpClient{
         return OkHttpClient.Builder()
@@ -33,7 +29,6 @@ object NetworkModule {
                 level = HttpLoggingInterceptor.Level.BODY
             }).build()
     }
-
     @Provides
     fun providesRetrofit(okHttpClient: OkHttpClient, moshi: Moshi): Retrofit{
 
@@ -43,12 +38,10 @@ object NetworkModule {
             .addConverterFactory(MoshiConverterFactory.create(moshi).asLenient())
             .build()
     }
-
     @Provides
     fun providesApi(retrofit: Retrofit): IDragonBallAPI {
         return retrofit.create(IDragonBallAPI::class.java)
     }
-
     @Provides
     fun providesApiLogin(retrofit: Retrofit): ILoginAPI {
         return  retrofit.create(ILoginAPI::class.java)
