+++ app/src/main/java/com/example/entregaandroidavanzadov2/utils/RepositoriesModuel.kt
@@ -12,11 +12,9 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoriesModule {
-
     @Binds
     abstract fun bindsRepository(repositoryImpl: RepositoryImpl): Repository
